+++ src/gql/base/ElementResolver.php
@@ -125,9 +125,14 @@ abstract class ElementResolver extends Resolver
         // Apply max result config
         $maxGraphqlResults = Craft::$app->getConfig()->getGeneral()->maxGraphqlResults;
 
+        // Reset negative limit to zero
+        if ((int)$query->limit < 0) {
+            $query->limit(0);
+        }
+
         if ($maxGraphqlResults > 0) {
-            // No sneaking around with negative limits!
-            $query->limit(min(max((int)$query->limit, 0), $maxGraphqlResults));
+            $queryLimit = is_null($query->limit) ? $maxGraphqlResults : min($maxGraphqlResults, $query->limit);
+            $query->limit($queryLimit);
         }
 
         return $query;