+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fixed a bug where Craft could attempt to use the Imagick driver even if it wasn’t available.
 - Fixed a PHP error that occurred if a test called `craft\test\fixtures\elements\ElementFixture::getModel()`. ([#7089](https://github.com/craftcms/cms/issues/7089))
 - Fixed a bug where `1x` image transforms weren’t using the original image transform.
+- Fixed a bug where sometimes Craft could not process an uploaded file if its extension was in uppercase.
 
 ## 3.5.15.1 - 2020-11-04